interpolated undefined gtfs's stop_times
related to https://github.com/CanalTP/navitia_model/issues/204

implement the stoptime's interpolation defined in the
[documentation](https://github.com/CanalTP/navitia_model/blob/master/src/documentation/gtfs_read.md#reading-stop_timestxt)
(a simple ventilation of the unknown stops).

Note: I'm not proud of the `Time::undefined`, but I don't see an
efficient way to wait for the whole file to be read. (Since the order of
the stop_times is not defined, we need to read the whole file, and for
the moment use thus create an `object::StopTime`, and modify it
afterward)